Refactor to use getter instead of computed value for logs
