+++ templates/vue/src/components/node-modal/QuizModal.vue
@@ -109,6 +109,7 @@ export default {
           id: Helpers.createUUID(),
           text: "",
           answers: { ...defaultQuestion.answers },
+          completed: false,
         },
       ]
     },
