+++ templates/vue/src/components/tyde/TydeBackpack.vue
@@ -1,14 +1,13 @@
 <template>
-    <div>
-        <input type="image" id="tyde-backpack-icon" :src="backpackUrl" @click="toggleMenu"/>
-        <tyde-menu
-        v-if="isMenuOpen"
-        :logs="logs"
-        @audio-change="toggleAudio"
-        @continue="continueTapestry"
-        @return-to-map="returnToMap"
-        />
-    </div>
+  <div>
+    <input type="image" id="tyde-backpack-icon" :src="backpackUrl" @click="toggleMenu"/>
+    <tyde-menu
+      v-if="isMenuOpen"
+      @audio-change="toggleAudio"
+      @continue="continueTapestry"
+      @return-to-map="returnToMap"
+    />
+  </div>
 </template>
 
 <script>
@@ -48,38 +47,10 @@ export default {
     this.backgroundAudio.loop = true
   },
   computed: {
-    ...mapGetters(["lightbox", "nodes"]),
+    ...mapGetters(["lightbox"]),
     backpackUrl() {
       return `${wpData.vue_uri}/${BackpackIcon.split("dist")[1]}`
     },
-    logs() {
-      const completedContents = this.nodes.filter(node => node.completed).map(node => {
-        return {
-          type: "content",
-          title: node.title,
-          imageURL: node.imageURL,
-          description: node.description,
-          isFavourite: node.isFavourite,
-        }
-      })
-      const completedActivities = this.nodes.reduce((activities, currentNode) => {
-        if (currentNode.quiz) {
-          const completedQuestions = currentNode.quiz.filter(q => q.completed).map(q => {
-            const keys = Object.keys(q.answers)
-            return keys.filter(key => q.answers[key]).map(key => {
-              return {
-                type: "activity",
-                title: q.text,
-                nodeId: currentNode.id,
-                [key]: q.answers[key],
-              }
-            })
-          })
-          return [...activities, ...completedQuestions.flat()]
-        }
-      }, [])
-      return [...completedContents, ...completedActivities]
-    }
   },
   methods: {
     ...mapMutations(["closeLightbox"]),
